requirements.txt
# Breathing Exercise Chatbot Requirements

# Agent framework
smolagents>=1.0.0

# Vector database and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Text processing
langchain-text-splitters>=0.0.1

# Environment variables
python-dotenv>=1.0.0

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
